Deprecate now-broken partial scene loading functions.
Note that these are used in some different demos owned by different people and it's hard for me to decide how to update these demos to still load a small subset of the scene without the context of the demos. As a result, we raise an error and leave it to the owners to replace the call.