Allow resetting to state dumps after initialization.
If provided pre-initialization, dumps can be used to set properties of states like particle systems that can only be set pre-initialization (e.g. sizes, colors).

If post-initialization, the pre-initialization states in the dump all need to match the states in the dump. Everything else (e.g. positions) will be reset to the dumped state.